// Frontend/src/types/room/type_RoomListings.ts
export interface IRoomListings {
    id: number;
    name: string;
    thumbnail: string;
    currency: string;
    category: string;
    price: number;
    priceType: string;
    bedroomCount: number;
    bathroomCount: number;
    bedCount: number;
    status: true;
    createdDate: string;
    updatedDate: string;
    location: string;
}